redirect to ems display of models after create and cancel
redirect back to the ems display view of the model, if that's where the user came from (= there's a storageManagerId).

this branch fixes this for host initiators, host initiator groups and volume mappings.
it also fixes for physical storages but that requires rebasing after recent changes in physical storage form are merged.

services and volumes are fixed in on-going PRs for other changes in these forms.